+++ src/natcap/invest/crop_production_regression.py
@@ -13,28 +13,29 @@ from .spec_utils import u
 from . import spec_utils
 from . import utils
 from . import validation
-from . import MODEL_METADATA
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
 
 LOGGER = logging.getLogger(__name__)
 
 CROPS = {
-    "barley": {"description": _("barley")},
-    "maize": {"description": _("maize")},
-    "oilpalm": {"description": _("oil palm")},
-    "potato": {"description": _("potato")},
-    "rice": {"description": _("rice")},
-    "soybean": {"description": _("soybean")},
-    "sugarbeet": {"description": _("sugar beet")},
-    "sugarcane": {"description": _("sugarcane")},
-    "wheat": {"description": _("wheat")}
+    "barley": {"description": gettext("barley")},
+    "maize": {"description": gettext("maize")},
+    "oilpalm": {"description": gettext("oil palm")},
+    "potato": {"description": gettext("potato")},
+    "rice": {"description": gettext("rice")},
+    "soybean": {"description": gettext("soybean")},
+    "sugarbeet": {"description": gettext("sugar beet")},
+    "sugarcane": {"description": gettext("sugarcane")},
+    "wheat": {"description": gettext("wheat")}
 }
 
 
 ARGS_SPEC = {
     "model_name": MODEL_METADATA["crop_production_regression"].model_title,
     "pyname": MODEL_METADATA["crop_production_regression"].pyname,
-    "userguide_html": MODEL_METADATA["crop_production_regression"].userguide,
+    "userguide": MODEL_METADATA["crop_production_regression"].userguide,
     "args_with_spatial_overlap": {
         "spatial_keys": ["landcover_raster_path", "aggregate_polygon_path"],
         "different_projections_ok": True,
@@ -57,11 +58,11 @@ ARGS_SPEC = {
                     "options": CROPS
                 }
             },
-            "about": _(
+            "about": gettext(
                 "A table that maps each LULC code from the LULC map to one of "
                 "the 10 canonical crop names representing the crop grown in "
                 "that LULC class."),
-            "name": _("LULC to crop table")
+            "name": gettext("LULC to crop table")
         },
         "fertilization_rate_table_path": {
             "type": "csv",
@@ -69,7 +70,7 @@ ARGS_SPEC = {
                 "crop_name": {
                     "type": "option_string",
                     "options": CROPS,
-                    "about": _("One of the supported crop types.")
+                    "about": gettext("One of the supported crop types.")
                 },
                 **{f"{nutrient}_rate": {
                     "type": "number",
@@ -77,9 +78,9 @@ ARGS_SPEC = {
                     "about": f"Rate of {nutrient} application for the crop."
                 } for nutrient in ["nitrogen", "phosphorus", "potassium"]}
             },
-            "about": _(
+            "about": gettext(
                 "A table that maps crops to fertilizer application rates."),
-            "name": _("fertilization rate table")
+            "name": gettext("fertilization rate table")
         },
         "aggregate_polygon_path": {
             **spec_utils.AOI,
@@ -152,8 +153,8 @@ ARGS_SPEC = {
                     }
                 }
             },
-            "about": _("The Crop Production datasets provided with the model."),
-            "name": _("model data")
+            "about": gettext("The Crop Production datasets provided with the model."),
+            "name": gettext("model data")
         }
     }
 }
@@ -823,7 +824,7 @@ def tabulate_regression_results(
 
                 # make a valid mask showing which pixels are not nodata
                 # if nodata value undefined, assume all pixels are valid
-                valid_mask = slice(None)
+                valid_mask = numpy.full(yield_block.shape, True)
                 if observed_yield_nodata is not None:
                     valid_mask = ~utils.array_equals_nodata(
                         yield_block, observed_yield_nodata)