+++ HISTORY.rst
@@ -39,13 +39,22 @@ Unreleased Changes
       ``universal2`` wheel architecture offered by ``scipy>=1.8.0``.
     * Fixing a small bug where drag-and-drop events in the Qt UI were not being
       handled correctly and were being ignored by the UI.
+    * Expose taskgraph logging level for the cli with
+      ``--taskgraph-log-level``.
+    * Fixed bug in validation of ``results_suffix`` so that special characters
+      like path separators, etc, are not allowed.
 * RouteDEM
     * Rename the arg ``calculate_downstream_distance`` to
       ``calculate_downslope_distance``. This is meant to clarify that it
       applies to pixels that are not part of a stream.
+* Crop Production
+    * Fixed a bug in both crop production models where the model would error if
+      an observed yield raster had no nodata value.
 * SDR
     * Fixed an issue with SDR where ``f.tif`` might not be recalculated if the
       file is modified or deleted after execution.
+    * Added basic type-checking for the ``lucode`` column of the biophysical
+      table. This avoids cryptic numpy errors later in runtime.
 
 3.10.2 (2022-02-08)
 -------------------